+++ npm-debug.log
@@ -1,0 +1,39 @@
+0 info it worked if it ends with ok
+1 verbose cli [ 'C:\\Program Files\\nodejs\\node.exe',
+1 verbose cli   'C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js',
+1 verbose cli   'start' ]
+2 info using npm@2.14.7
+3 info using node@v4.2.3
+4 verbose run-script [ 'prestart', 'start', 'poststart' ]
+5 info prestart morse-server@1.0.0
+6 info start morse-server@1.0.0
+7 verbose unsafe-perm in lifecycle true
+8 info morse-server@1.0.0 Failed to exec start script
+9 verbose stack Error: morse-server@1.0.0 start: `node index.js`
+9 verbose stack Exit status 1
+9 verbose stack     at EventEmitter.<anonymous> (C:\Users\Csaba\AppData\Roaming\nvm\v4.2.3\node_modules\npm\lib\utils\lifecycle.js:214:16)
+9 verbose stack     at emitTwo (events.js:87:13)
+9 verbose stack     at EventEmitter.emit (events.js:172:7)
+9 verbose stack     at ChildProcess.<anonymous> (C:\Users\Csaba\AppData\Roaming\nvm\v4.2.3\node_modules\npm\lib\utils\spawn.js:24:14)
+9 verbose stack     at emitTwo (events.js:87:13)
+9 verbose stack     at ChildProcess.emit (events.js:172:7)
+9 verbose stack     at maybeClose (internal/child_process.js:818:16)
+9 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:211:5)
+10 verbose pkgid morse-server@1.0.0
+11 verbose cwd H:\Programming\projects\electron\morse-server
+12 error Windows_NT 6.3.9600
+13 error argv "C:\\Program Files\\nodejs\\node.exe" "C:\\Program Files\\nodejs\\node_modules\\npm\\bin\\npm-cli.js" "start"
+14 error node v4.2.3
+15 error npm  v2.14.7
+16 error code ELIFECYCLE
+17 error morse-server@1.0.0 start: `node index.js`
+17 error Exit status 1
+18 error Failed at the morse-server@1.0.0 start script 'node index.js'.
+18 error This is most likely a problem with the morse-server package,
+18 error not with npm itself.
+18 error Tell the author that this fails on your system:
+18 error     node index.js
+18 error You can get their info via:
+18 error     npm owner ls morse-server
+18 error There is likely additional logging output above.
+19 verbose exit [ 1, true ]
